Move trait descriptions to @scripts/config/traits.ts
Also add "relic" as equipment trait
