+++ src/scripts/config/index.ts
@@ -27,7 +27,6 @@ import {
     WeaponPF2e,
 } from "@item";
 import { ConditionSlug } from "@item/condition/types.ts";
-import { RANGE_TRAITS } from "@item/data/values.ts";
 import { DeityDomain } from "@item/deity/types.ts";
 import { FeatCategory } from "@item/feat/index.ts";
 import { WEAPON_PROPERTY_RUNES } from "@item/physical/runes.ts";
@@ -62,6 +61,7 @@ import {
     preciousMaterials,
     spellOtherTraits,
     spellTraits,
+    traitDescriptions,
     vehicleTraits,
     weaponTraits,
 } from "./traits.ts";
@@ -196,436 +196,6 @@ const equivalentWeapons = {
     "big-boom-gun": "hand-cannon",
     "spoon-gun": "hand-cannon",
 } as const;
-
-const rangeDescriptions = RANGE_TRAITS.reduce(
-    (descriptions, trait) => mergeObject(descriptions, { [trait]: "PF2E.TraitDescriptionRange" }),
-    {} as Record<(typeof RANGE_TRAITS)[number], string>
-);
-
-const preciousMaterialDescriptions = {
-    abysium: "PF2E.PreciousMaterialAbysiumDescription",
-    adamantine: "PF2E.PreciousMaterialAdamantineDescription",
-    "cold-iron": "PF2E.PreciousMaterialColdIronDescription",
-    darkwood: "PF2E.PreciousMaterialDarkwoodDescription",
-    djezet: "PF2E.PreciousMaterialDjezetDescription",
-    dragonhide: "PF2E.PreciousMaterialDragonhideDescription",
-    "grisantian-pelt": "PF2E.PreciousMaterialGrisantianPeltDescription",
-    inubrix: "PF2E.PreciousMaterialInubrixDescription",
-    mithral: "PF2E.PreciousMaterialMithralDescription",
-    noqual: "PF2E.PreciousMaterialNoqualDescription",
-    orichalcum: "PF2E.PreciousMaterialOrichalcumDescription",
-    siccatite: "PF2E.PreciousMaterialSiccatiteDescription",
-    silver: "PF2E.PreciousMaterialSilverDescription",
-    "sisterstone-dusk": "PF2E.PreciousMaterialSisterstoneDescription",
-    "sisterstone-scarlet": "PF2E.PreciousMaterialSisterstoneDescription",
-    "sovereign-steel": "PF2E.PreciousMaterialSovereignSteelDescription",
-    warpglass: "PF2E.PreciousMaterialWarpglassDescription",
-};
-
-const traitsDescriptions = {
-    aasimar: "PF2E.TraitDescriptionAasimar",
-    aberration: "PF2E.TraitDescriptionAberration",
-    abjuration: "PF2E.TraitDescriptionAbjuration",
-    acid: "PF2E.TraitDescriptionAcid",
-    additive0: "PF2E.TraitDescriptionAdditive",
-    additive1: "PF2E.TraitDescriptionAdditive",
-    additive2: "PF2E.TraitDescriptionAdditive",
-    additive3: "PF2E.TraitDescriptionAdditive",
-    adjusted: "PF2E.TraitDescriptionAdjusted",
-    adjustment: "PF2E.TraitDescriptionAdjustment",
-    aeon: "PF2E.TraitDescriptionAeon",
-    aesir: "PF2E.TraitDescriptionAesir",
-    aftermath: "PF2E.TraitDescriptionAftermath",
-    agathion: "PF2E.TraitDescriptionAgathion",
-    agile: "PF2E.TraitDescriptionAgile",
-    air: "PF2E.TraitDescriptionAir",
-    alchemical: "PF2E.TraitDescriptionAlchemical",
-    alchemist: "PF2E.TraitDescriptionAlchemist",
-    amphibious: "PF2E.TraitDescriptionAmphibious",
-    anadi: "PF2E.TraitDescriptionAnadi",
-    android: "PF2E.TraitDescriptionAndroid",
-    apex: "PF2E.TraitDescriptionApex",
-    aphorite: "PF2E.TraitDescriptionAphorite",
-    aquadynamic: "PF2E.TraitDescriptionAquadynamic",
-    arcane: "PF2E.TraitDescriptionArcane",
-    archetype: "PF2E.TraitDescriptionArchetype",
-    artifact: "PF2E.TraitDescriptionArtifact",
-    attached: "PF2E.TraitDescriptionAttached",
-    "attached-to-crossbow-or-firearm": "PF2E.TraitDescriptionAttached",
-    "attached-to-shield": "PF2E.TraitDescriptionAttached",
-    attack: "PF2E.TraitDescriptionAttack",
-    auditory: "PF2E.TraitDescriptionAuditory",
-    aura: "PF2E.TraitDescriptionAura",
-    automaton: "PF2E.TraitDescriptionAutomaton",
-    azarketi: "PF2E.TraitDescriptionAzarketi",
-    backstabber: "PF2E.TraitDescriptionBackstabber",
-    backswing: "PF2E.TraitDescriptionBackswing",
-    barbarian: "PF2E.TraitDescriptionBarbarian",
-    bard: "PF2E.TraitDescriptionBard",
-    beastkin: "PF2E.TraitDescriptionBeastkin",
-    bomb: "PF2E.TraitDescriptionBomb",
-    brace: "PF2E.TraitDescriptionBrace",
-    brutal: "PF2E.TraitDescriptionBrutal",
-    bulwark: "PF2E.TraitDescriptionBulwark",
-    cantrip: "PF2E.TraitDescriptionCantrip",
-    "capacity-2": "PF2E.TraitDescriptionCapacity",
-    "capacity-3": "PF2E.TraitDescriptionCapacity",
-    "capacity-4": "PF2E.TraitDescriptionCapacity",
-    "capacity-5": "PF2E.TraitDescriptionCapacity",
-    catalyst: "PF2E.TraitDescriptionCatalyst",
-    catfolk: "PF2E.TraitDescriptionCatfolk",
-    champion: "PF2E.TraitDescriptionChampion",
-    changeling: "PF2E.TraitDescriptionChangeling",
-    chaotic: "PF2E.TraitDescriptionChaotic",
-    class: "PF2E.TraitDescriptionClass",
-    cleric: "PF2E.TraitDescriptionCleric",
-    climbing: "PF2E.TraitDescriptionClimbing",
-    clockwork: "PF2E.TraitDescriptionClockwork",
-    cobbled: "PF2E.TraitDescriptionCobbled",
-    coda: "PF2E.TraitDescriptionCoda",
-    cold: "PF2E.TraitDescriptionCold",
-    combination: "PF2E.TraitDescriptionCombination",
-    comfort: "PF2E.TraitDescriptionComfort",
-    common: "PF2E.TraitDescriptionCommon",
-    companion: "PF2E.TraitDescriptionCompanion",
-    complex: "PF2E.TraitDescriptionComplex",
-    composition: "PF2E.TraitDescriptionComposition",
-    concealable: "PF2E.TraitDescriptionConcealable",
-    concentrate: "PF2E.TraitDescriptionConcentrate",
-    concussive: "PF2E.TraitDescriptionConcussive",
-    conjuration: "PF2E.TraitDescriptionConjuration",
-    conrasu: "PF2E.TraitDescriptionConrasu",
-    consecration: "PF2E.TraitDescriptionConsecration",
-    consumable: "PF2E.TraitDescriptionConsumable",
-    contact: "PF2E.TraitDescriptionContact",
-    contingency: "PF2E.TraitDescriptionContingency",
-    contract: "PF2E.TraitDescriptionContract",
-    "critical-fusion": "PF2E.TraitDescriptionCriticalFusion",
-    curse: "PF2E.TraitDescriptionCurse",
-    cursebound: "PF2E.TraitDescriptionCursebound",
-    cursed: "PF2E.TraitDescriptionCursed",
-    darkness: "PF2E.TraitDescriptionDarkness",
-    darvakka: "PF2E.TraitDescriptionDarvakka",
-    "deadly-2d10": "PF2E.TraitDescriptionDeadly",
-    "deadly-2d12": "PF2E.TraitDescriptionDeadly",
-    "deadly-2d8": "PF2E.TraitDescriptionDeadly",
-    "deadly-3d10": "PF2E.TraitDescriptionDeadly",
-    "deadly-3d12": "PF2E.TraitDescriptionDeadly",
-    "deadly-3d8": "PF2E.TraitDescriptionDeadly",
-    "deadly-4d10": "PF2E.TraitDescriptionDeadly",
-    "deadly-4d12": "PF2E.TraitDescriptionDeadly",
-    "deadly-4d8": "PF2E.TraitDescriptionDeadly",
-    "deadly-d10": "PF2E.TraitDescriptionDeadly",
-    "deadly-d12": "PF2E.TraitDescriptionDeadly",
-    "deadly-d6": "PF2E.TraitDescriptionDeadly",
-    "deadly-d8": "PF2E.TraitDescriptionDeadly",
-    "deadly-d4": "PF2E.TraitDescriptionDeadly",
-    death: "PF2E.TraitDescriptionDeath",
-    "deflecting-bludgeoning": "PF2E.TraitDescriptionDeflecting",
-    "deflecting-physical-ranged": "PF2E.TraitDescriptionDeflecting",
-    "deflecting-slashing": "PF2E.TraitDescriptionDeflecting",
-    dedication: "PF2E.TraitDescriptionDedication",
-    detection: "PF2E.TraitDescriptionDetection",
-    deviant: "PF2E.TraitDescriptionDeviant",
-    dhampir: "PF2E.TraitDescriptionDhampir",
-    disarm: "PF2E.TraitDescriptionDisarm",
-    disease: "PF2E.TraitDescriptionDisease",
-    divination: "PF2E.TraitDescriptionDivination",
-    divine: "PF2E.TraitDescriptionDivine",
-    "double-barrel": "PF2E.TraitDescriptionDoubleBarrel",
-    downtime: "PF2E.TraitDescriptionDowntime",
-    drug: "PF2E.TraitDescriptionDrug",
-    druid: "PF2E.TraitDescriptionDruid",
-    duskwalker: "PF2E.TraitDescriptionDuskwalker",
-    dwarf: "PF2E.TraitDescriptionDwarf",
-    earth: "PF2E.TraitDescriptionEarth",
-    eidolon: "PF2E.TraitDescriptionEidolon",
-    electricity: "PF2E.TraitDescriptionElectricity",
-    elf: "PF2E.TraitDescriptionElf",
-    elixir: "PF2E.TraitDescriptionElixir",
-    emotion: "PF2E.TraitDescriptionEmotion",
-    enchantment: "PF2E.TraitDescriptionEnchantment",
-    "entrench-melee": "PF2E.TraitDescriptionEntrenchMelee",
-    "entrench-ranged": "PF2E.TraitDescriptionEntrenchRanged",
-    environment: "PF2E.TraitDescriptionEnvironment",
-    esoterica: "PF2E.TraitDescriptionEsoterica",
-    evil: "PF2E.TraitDescriptionEvil",
-    evocation: "PF2E.TraitDescriptionEvocation",
-    evolution: "PF2E.TraitDescriptionEvolution",
-    expandable: "PF2E.TraitDescriptionExpandable",
-    exploration: "PF2E.TraitDescriptionExploration",
-    extradimensional: "PF2E.TraitDescriptionExtradimensional",
-    "fatal-aim-d10": "PF2E.TraitDescriptionFatalAim",
-    "fatal-aim-d12": "PF2E.TraitDescriptionFatalAim",
-    "fatal-d10": "PF2E.TraitDescriptionFatal",
-    "fatal-d12": "PF2E.TraitDescriptionFatal",
-    "fatal-d8": "PF2E.TraitDescriptionFatal",
-    fear: "PF2E.TraitDescriptionFear",
-    fetchling: "PF2E.TraitDescriptionFetchling",
-    fey: "PF2E.TraitDescriptionFey",
-    fighter: "PF2E.TraitDescriptionFighter",
-    finesse: "PF2E.TraitDescriptionFinesse",
-    finisher: "PF2E.TraitDescriptionFinisher",
-    fire: "PF2E.TraitDescriptionFire",
-    fleshwarp: "PF2E.TraitDescriptionFleshwarp",
-    flexible: "PF2E.TraitDescriptionFlexible",
-    flourish: "PF2E.TraitDescriptionFlourish",
-    focused: "PF2E.TraitDescriptionFocused",
-    foldaway: "PF2E.TraitDescriptionFoldaway",
-    force: "PF2E.TraitDescriptionForce",
-    forceful: "PF2E.TraitDescriptionForceful",
-    fortune: "PF2E.TraitDescriptionFortune",
-    "free-hand": "PF2E.TraitDescriptionFreeHand",
-    fulu: "PF2E.TraitDescriptionFulu",
-    fungus: "PF2E.TraitDescriptionFungus",
-    gadget: "PF2E.TraitDescriptionGadget",
-    ganzi: "PF2E.TraitDescriptionGanzi",
-    general: "PF2E.TraitDescriptionGeneral",
-    genie: "PF2E.TraitDescriptionGenie",
-    geniekin: "PF2E.TraitDescriptionGeniekin",
-    ghoran: "PF2E.TraitDescriptionGhoran",
-    ghul: "PF2E.TraitDescriptionGhul",
-    gnoll: "PF2E.TraitDescriptionGnoll",
-    gnome: "PF2E.TraitDescriptionGnome",
-    goblin: "PF2E.TraitDescriptionGoblin",
-    goloma: "PF2E.TraitDescriptionGoloma",
-    good: "PF2E.TraitDescriptionGood",
-    grapple: "PF2E.TraitDescriptionGrapple",
-    grimoire: "PF2E.TraitDescriptionGrimoire",
-    grippli: "PF2E.TraitDescriptionGrippli",
-    gunslinger: "PF2E.TraitDescriptionGunslinger",
-    halfling: "PF2E.TraitDescriptionHalfling",
-    "half-elf": "PF2E.TraitDescriptionHalfElf",
-    "half-orc": "PF2E.TraitDescriptionHalfOrc",
-    "jousting-d6": "PF2E.TraitDescriptionJousting",
-    hampering: "PF2E.TraitDescriptionHampering",
-    harnessed: "PF2E.TraitDescriptionHarnessed",
-    haunt: "PF2E.TraitDescriptionHaunt",
-    healing: "PF2E.TraitDescriptionHealing",
-    "hefty-14": "PF2E.TraitDescriptionHefty",
-    hex: "PF2E.TraitDescriptionHex",
-    hindering: "PF2E.TraitDescriptionHindering",
-    hobgoblin: "PF2E.TraitDescriptionHobgoblin",
-    human: "PF2E.TraitDescriptionHuman",
-    humanoid: "PF2E.TraitDescriptionHumanoid",
-    ifrit: "PF2E.TraitDescriptionIfrit",
-    illusion: "PF2E.TraitDescriptionIllusion",
-    incapacitation: "PF2E.TraitDescriptionIncapacitation",
-    incarnate: "PF2E.TraitDescriptionIncarnate",
-    incorporeal: "PF2E.TraitDescriptionIncorporeal",
-    infused: "PF2E.TraitDescriptionInfused",
-    ingested: "PF2E.TraitDescriptionIngested",
-    inhaled: "PF2E.TraitDescriptionInhaled",
-    injection: "PF2E.TraitDescriptionInjection",
-    injury: "PF2E.TraitDescriptionInjury",
-    inscribed: "PF2E.TraitDescriptionInscribed",
-    "integrated-1d6-b": "PF2E.TraitDescriptionIntegrated",
-    "integrated-1d6-p": "PF2E.TraitDescriptionIntegrated",
-    "integrated-1d6-s": "PF2E.TraitDescriptionIntegrated",
-    "integrated-1d6-s-versatile-p": "PF2E.TraitDescriptionIntegrated",
-    intelligent: "PF2E.TraitDescriptionIntelligent",
-    inventor: "PF2E.TraitDescriptionInventor",
-    invested: "PF2E.TraitDescriptionInvested",
-    investigator: "PF2E.TraitDescriptionInvestigator",
-    kashrishi: "PF2E.TraitDescriptionKashrishi",
-    kickback: "PF2E.TraitDescriptionKickback",
-    kitsune: "PF2E.TraitDescriptionKitsune",
-    kobold: "PF2E.TraitDescriptionKobold",
-    laminar: "PF2E.TraitDescriptionLaminar",
-    "launching-dart": "PF2E.TraitDescriptionLaunching",
-    lawful: "PF2E.TraitDescriptionLawful",
-    leshy: "PF2E.TraitDescriptionLeshy",
-    light: "PF2E.TraitDescriptionLight",
-    linguistic: "PF2E.TraitDescriptionLinguistic",
-    litany: "PF2E.TraitDescriptionLitany",
-    lizardfolk: "PF2E.TraitDescriptionLizardfolk",
-    lozenge: "PF2E.TraitDescriptionLozenge",
-    magical: "PF2E.TraitDescriptionMagical",
-    magus: "PF2E.TraitDescriptionMagus",
-    manipulate: "PF2E.TraitDescriptionManipulate",
-    mechanical: "PF2E.TraitDescriptionMechanical",
-    mental: "PF2E.TraitDescriptionMental",
-    metal: "PF2E.TraitDescriptionMetal",
-    metamagic: "PF2E.TraitDescriptionMetamagic",
-    mindless: "PF2E.TraitDescriptionMindless",
-    mindshift: "PF2E.TraitDescriptionMindshift",
-    minion: "PF2E.TraitDescriptionMinion",
-    misfortune: "PF2E.TraitDescriptionMisfortune",
-    missive: "PF2E.TraitDescriptionMissive",
-    modification: "PF2E.TraitDescriptionModification",
-    modular: "PF2E.TraitDescriptionModular",
-    monk: "PF2E.TraitDescriptionMonk",
-    morph: "PF2E.TraitDescriptionMorph",
-    mounted: "PF2E.TraitDescriptionMounted",
-    move: "PF2E.TraitDescriptionMove",
-    multiclass: "PF2E.TraitDescriptionMulticlass",
-    mutagen: "PF2E.TraitDescriptionMutagen",
-    nagaji: "PF2E.TraitDescriptionNagaji",
-    necromancy: "PF2E.TraitDescriptionNecromancy",
-    negative: "PF2E.TraitDescriptionNegative",
-    noisy: "PF2E.TraitDescriptionNoisy",
-    nonlethal: "PF2E.TraitDescriptionNonlethal",
-    oath: "PF2E.TraitDescriptionOath",
-    occult: "PF2E.TraitDescriptionOccult",
-    oil: "PF2E.TraitDescriptionOil",
-    olfactory: "PF2E.TraitDescriptionOlfactory",
-    open: "PF2E.TraitDescriptionOpen",
-    oracle: "PF2E.TraitDescriptionOracle",
-    orc: "PF2E.TraitDescriptionOrc",
-    oread: "PF2E.TraitDescriptionOread",
-    parry: "PF2E.TraitDescriptionParry",
-    peachwood: "PF2E.PreciousMaterialPeachwoodDescription",
-    plant: "PF2E.TraitDescriptionPlant",
-    poison: "PF2E.TraitDescriptionPoison",
-    polymorph: "PF2E.TraitDescriptionPolymorph",
-    ponderous: "PF2E.TraitDescriptionPonderous",
-    poppet: "PF2E.TraitDescriptionPoppet",
-    portable: "PF2E.TraitDescriptionPortable",
-    positive: "PF2E.TraitDescriptionPositive",
-    possession: "PF2E.TraitDescriptionPossession",
-    potion: "PF2E.TraitDescriptionPotion",
-    precious: "PF2E.TraitDescriptionPrecious",
-    prediction: "PF2E.TraitDescriptionPrediction",
-    press: "PF2E.TraitDescriptionPress",
-    primal: "PF2E.TraitDescriptionPrimal",
-    propulsive: "PF2E.TraitDescriptionPropulsive",
-    psyche: "PF2E.TraitDescriptionPsyche",
-    psychic: "PF2E.TraitDescriptionPsychic",
-    rage: "PF2E.TraitDescriptionRage",
-    ranger: "PF2E.TraitDescriptionRanger",
-    ...rangeDescriptions,
-    "ranged-trip": "PF2E.TraitDescriptionRangedTrip",
-    rare: "PF2E.TraitDescriptionRare",
-    ratfolk: "PF2E.TraitDescriptionRatfolk",
-    razing: "PF2E.TraitDescriptionRazing",
-    reach: "PF2E.TraitDescriptionReach",
-    "reach-0": "PF2E.TraitDescriptionReach",
-    "reach-10": "PF2E.TraitDescriptionReach",
-    "reach-100": "PF2E.TraitDescriptionReach",
-    "reach-1000": "PF2E.TraitDescriptionReach",
-    "reach-15": "PF2E.TraitDescriptionReach",
-    "reach-20": "PF2E.TraitDescriptionReach",
-    "reach-25": "PF2E.TraitDescriptionReach",
-    "reach-30": "PF2E.TraitDescriptionReach",
-    "reach-40": "PF2E.TraitDescriptionReach",
-    "reach-50": "PF2E.TraitDescriptionReach",
-    "reach-60": "PF2E.TraitDescriptionReach",
-    recovery: "PF2E.TraitDescriptionRecovery",
-    reflection: "PF2E.TraitDescriptionReflection",
-    reload: "PF2E.TraitDescriptionReload",
-    "reload-0": "PF2E.TraitDescriptionReload",
-    "reload-1": "PF2E.TraitDescriptionReload",
-    "reload-1-min": "PF2E.TraitDescriptionReload",
-    "reload-2": "PF2E.TraitDescriptionReload",
-    repeating: "PF2E.TraitDescriptionRepeating",
-    resonant: "PF2E.TraitDescriptionResonant",
-    revelation: "PF2E.TraitDescriptionRevelation",
-    rogue: "PF2E.TraitDescriptionRogue",
-    saggorak: "PF2E.TraitDescriptionSaggorak",
-    "scatter-10": "PF2E.TraitDescriptionScatter",
-    "scatter-15": "PF2E.TraitDescriptionScatter",
-    "scatter-5": "PF2E.TraitDescriptionScatter",
-    scroll: "PF2E.TraitDescriptionScroll",
-    scrying: "PF2E.TraitDescriptionScrying",
-    secret: "PF2E.TraitDescriptionSecret",
-    shadow: "PF2E.TraitDescriptionShadow",
-    "shield-throw-20": "PF2E.TraitDescriptionShieldThrow",
-    "shield-throw-30": "PF2E.TraitDescriptionShieldThrow",
-    shisk: "PF2E.TraitDescriptionShisk",
-    shoony: "PF2E.TraitDescriptionShoony",
-    shove: "PF2E.TraitDescriptionShove",
-    skeleton: "PF2E.TraitDescriptionSkeleton",
-    skill: "PF2E.TraitDescriptionSkill",
-    sleep: "PF2E.TraitDescriptionSleep",
-    snare: "PF2E.TraitDescriptionSnare",
-    social: "PF2E.TraitDescriptionSocial",
-    sonic: "PF2E.TraitDescriptionSonic",
-    sorcerer: "PF2E.TraitDescriptionSorcerer",
-    spellgun: "PF2E.TraitDescriptionSpellgun",
-    spellheart: "PF2E.TraitDescriptionSpellheart",
-    spellshot: "PF2E.TraitNoDescription",
-    splash: "PF2E.TraitDescriptionSplash",
-    "splash-10": "PF2E.TraitDescriptionSplash10",
-    sprite: "PF2E.TraitDescriptionSprite",
-    staff: "PF2E.TraitDescriptionStaff",
-    stance: "PF2E.TraitDescriptionStance",
-    steam: "PF2E.TraitDescriptionSteam",
-    strix: "PF2E.TraitDescriptionStrix",
-    structure: "PF2E.TraitDescriptionStructure",
-    suli: "PF2E.TraitDescriptionSuli",
-    summon: "PF2E.TraitDescriptionSummon",
-    summoned: "PF2E.TraitDescriptionSummoned",
-    summoner: "PF2E.TraitDescriptionSummoner",
-    swashbuckler: "PF2E.TraitDescriptionSwashbuckler",
-    sweep: "PF2E.TraitDescriptionSweep",
-    sylph: "PF2E.TraitDescriptionSylph",
-    talisman: "PF2E.TraitDescriptionTalisman",
-    tandem: "PF2E.TraitDescriptionTandem",
-    tattoo: "PF2E.TraitDescriptionTattoo",
-    tech: "PF2E.TraitDescriptionTech",
-    telepathy: "PF2E.TraitDescriptionTelepathy",
-    teleportation: "PF2E.TraitDescriptionTeleportation",
-    tengu: "PF2E.TraitDescriptionTengu",
-    tethered: "PF2E.TraitDescriptionTethered",
-    thaumaturge: "PF2E.TraitDescriptionThaumaturge",
-    thrown: "PF2E.TraitDescriptionThrown",
-    "thrown-10": "PF2E.TraitDescriptionThrown",
-    "thrown-100": "PF2E.TraitDescriptionThrown",
-    "thrown-15": "PF2E.TraitDescriptionThrown",
-    "thrown-20": "PF2E.TraitDescriptionThrown",
-    "thrown-200": "PF2E.TraitDescriptionThrown",
-    "thrown-25": "PF2E.TraitDescriptionThrown",
-    "thrown-30": "PF2E.TraitDescriptionThrown",
-    "thrown-40": "PF2E.TraitDescriptionThrown",
-    "thrown-50": "PF2E.TraitDescriptionThrown",
-    "thrown-60": "PF2E.TraitDescriptionThrown",
-    "thrown-80": "PF2E.TraitDescriptionThrown",
-    tiefling: "PF2E.TraitDescriptionTiefling",
-    time: "PF2E.TraitDescriptionTime",
-    training: "PF2E.TraitDescriptionTraining",
-    transmutation: "PF2E.TraitDescriptionTransmutation",
-    trap: "PF2E.TraitDescriptionTrap",
-    trip: "PF2E.TraitDescriptionTrip",
-    "true-name": "PF2E.TraitDescriptionTrueName",
-    twin: "PF2E.TraitDescriptionTwin",
-    "two-hand-d10": "PF2E.TraitDescriptionTwoHand",
-    "two-hand-d12": "PF2E.TraitDescriptionTwoHand",
-    "two-hand-d6": "PF2E.TraitDescriptionTwoHand",
-    "two-hand-d8": "PF2E.TraitDescriptionTwoHand",
-    unarmed: "PF2E.TraitDescriptionUnarmed",
-    uncommon: "PF2E.TraitDescriptionUncommon",
-    undine: "PF2E.TraitDescriptionUndine",
-    unique: "PF2E.TraitDescriptionUnique",
-    unstable: "PF2E.TraitDescriptionUnstable",
-    vanara: "PF2E.TraitDescriptionVanara",
-    vishkanya: "PF2E.TraitDescriptionVishkanya",
-    vehicular: "PF2E.TraitDescriptionVehicular",
-    "versatile-acid": "PF2E.TraitDescriptionVersatile",
-    "versatile-b": "PF2E.TraitDescriptionVersatile",
-    "versatile-chaotic": "PF2E.TraitDescriptionVersatile",
-    "versatile-cold": "PF2E.TraitDescriptionVersatile",
-    "versatile-electricity": "PF2E.TraitDescriptionVersatile",
-    "versatile-evil": "PF2E.TraitDescriptionVersatile",
-    "versatile-fire": "PF2E.TraitDescriptionVersatile",
-    "versatile-force": "PF2E.TraitDescriptionVersatile",
-    "versatile-good": "PF2E.TraitDescriptionVersatile",
-    "versatile-lawful": "PF2E.TraitDescriptionVersatile",
-    "versatile-negative": "PF2E.TraitDescriptionVersatile",
-    "versatile-p": "PF2E.TraitDescriptionVersatile",
-    "versatile-positive": "PF2E.TraitDescriptionVersatile",
-    "versatile-s": "PF2E.TraitDescriptionVersatile",
-    "versatile-sonic": "PF2E.TraitDescriptionVersatile",
-    virulent: "PF2E.TraitDescriptionVirulent",
-    visual: "PF2E.TraitDescriptionVisual",
-    "volley-20": "PF2E.TraitDescriptionVolley",
-    "volley-30": "PF2E.TraitDescriptionVolley",
-    "volley-50": "PF2E.TraitDescriptionVolley",
-    wand: "PF2E.TraitDescriptionWand",
-    water: "PF2E.TraitDescriptionWater",
-    witch: "PF2E.TraitDescriptionWitch",
-    wizard: "PF2E.TraitDescriptionWizard",
-    ...preciousMaterialDescriptions,
-};
 
 const preciousMaterialGrades: Record<PreciousMaterialGrade, string> = {
     low: "PF2E.PreciousMaterialLowGrade",
@@ -1096,7 +666,7 @@ export const PF2ECONFIG = {
     hazardTraits,
     vehicleTraits,
 
-    traitsDescriptions,
+    traitsDescriptions: traitDescriptions,
 
     weaponHands: {
         1: "PF2E.WeaponHands1",
